requirements.txt
BeautifulSoup==3.2.1
Django==1.4.13
-e git://github.com/MichalMaM/django-commander.git@15ebe29728821dd986000bac178a65f6845e3e99#egg=Django_commander-dev
Fabric==1.6.1
MySQL-python==1.2.5
PyYAML==3.10
South==0.7.6
argparse==1.2.1
boto==2.2.1
chardet==2.1.1
dateutils==0.6.6
distribute==0.6.31
dj-database-url==0.2.1
django-admin-bootstrapped==0.3.2
django-csvimport==0.9
django-data-exports==0.3
django-debug-toolbar==0.9.4
django-inspect-model==0.4
django-mass-edit==1.2
django-redis==3.3
django-storages==1.1.4
django-tastypie==0.9.11
django-yamlfield==0.4
ez-setup==0.9
geopy==0.95.1
gunicorn==0.17.4
mechanize==0.2.5
mimeparse==0.1.3
paramiko==1.10.1
pycrypto==2.6
python-dateutil==1.5
python-ptrace==0.6.4
pytz==2012c
redis==2.8.0
requests==2.2.1
simplejson==3.3.0
titlecase==0.5.1
tweepy==2.1
wsgiref==0.1.2
yolk==0.4.3
